+++ divert/core.py
@@ -1,9 +1,11 @@
+from __future__ import annotations
+
 from typing import TypeVar, Generic, Any
 
 Payload = TypeVar("Payload")
 
 
-class Diversion(Exception):
+class Diversion(BaseException):
     """Diversion from the normal execution flow."""
 
     def __init__(self, jumps: int = 1) -> None:
@@ -15,6 +17,19 @@ class Diversion(Exception):
         """Whether the diversion should be raised again."""
         self._jumps_remaining -= 1
         return self._jumps_remaining > 0
+
+    @classmethod
+    def make_weak(cls, *args: Any, **kwargs: Any) -> Diversion:
+        """Make weak variant of the diversion class.
+
+        By dynamically forming it provides support for generic extensions of the base diversions.
+        The minor performance hit is acceptable vs. having to maintain `WeakDiversion` instances for each extension.
+        """
+
+        class WeakDiversion(cls, Exception):
+            """Weak variation of the diversion, so it can be caught by *normal* user exceptions."""
+
+        return WeakDiversion(*args, **kwargs)
 
 
 class PayloadDiversion(Diversion, Generic[Payload]):