+++ changelog.md
@@ -1,5 +1,10 @@
 # Flow Control
 
+## 0.2.0
+
+* feature: support for weak diversion variants
+  * a weak variant allows it to be caught by *normal* try-catches when using `except Exception:`
+
 ## 0.1.0
 
 * feature: basic flow edge and jump to edge